requirements.txt
farm-haystack>=1.20.0
anthropic>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.5
requests>=2.28.0
